LayoutTests/fast/forms/input-number-floor-ceiling-stepup.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('stepUp() Precision Non-integrals');

var input = document.createElement('input');
var changeEventCounter = 0;
input.onchange = function() { ++changeEventCounter; };

function setInputAttributes(min, max, step, value) {
    input.min = min;
    input.max = max;
    input.step = step;
    input.value = value;
}

function stepUpExplicitBounds(min, max, step, value, stepCount) {
    setInputAttributes(min, max, step, value);
    if (typeof stepCount !== 'undefined')
        input.stepUp(stepCount);
    else
        input.stepUp();
    return input.value;
}

var input = document.createElement('input');

input.type = 'number';
debug('StepUp Precision Non-integrals');
shouldBe('stepUpExplicitBounds("1.3", "", "0.2", "1.51", 1)', '"1.7"');
</script>
</body>
</html>
