LayoutTests/fast/forms/input-number-floor-ceiling-stepup-expected.txt

stepUp() Precision Non-integrals

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


StepUp Precision Non-integrals
PASS stepUpExplicitBounds("1.3", "", "0.2", "1.51", 1) is "1.7"
PASS successfullyParsed is true

TEST COMPLETE
